+++ src/components/common/Heading.jsx
@@ -1,7 +1,13 @@
-import React from 'react';
+import React, {Fragment} from 'react';
+import Notification from './Notification'
 
-const Heading = (props) => {
-    return ( <h1>{props.text}</h1> );
+const Heading = ({text,errors,notice}) => {
+    return ( 
+    <Fragment>
+        <h1>{text}</h1> 
+        {/* <Notification errors={errors} notice={notice} /> */}
+    </Fragment>
+    );
 }
  
 export default Heading;
